teachers_dir/requirement.txt
appdirs==1.4.3
asgiref==3.2.10
beautifulsoup4==4.9.0
bs4==0.0.1
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==3.0.7
django-import-export==2.2.0
et-xmlfile==1.0.1
filelock==3.0.12
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
pytz==2020.1
PyYAML==5.3.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tablib==2.0.0
virtualenv==20.0.8
virtualenvwrapper-win==1.2.6
xlrd==1.2.0
xlwt==1.3.0
